add primitive fire-period system
guns have fire periods, i.e. time between shots

I added some variables to the player to enable this, and changed the
"pick up this gun" functions to set everything related to the gun. The
gun itself doesn't actually know anything abt it, it just calls one of
the player's functions that sets the players values to what the gun
would have.